+++ cdn-20180510/alibabacloud_cdn20180510/models.py
@@ -30232,11 +30232,17 @@ class DisableRealtimeLogDeliveryRequest(TeaModel):
     def __init__(
         self,
         domain: str = None,
+        logstore: str = None,
+        project: str = None,
+        region: str = None,
     ):
         # The accelerated domain name for which you want to disable real-time log delivery. You can specify multiple domain names and separate them with commas (,).
         # 
         # This parameter is required.
         self.domain = domain
+        self.logstore = logstore
+        self.project = project
+        self.region = region
 
     def validate(self):
         pass
@@ -30249,12 +30255,24 @@ class DisableRealtimeLogDeliveryRequest(TeaModel):
         result = dict()
         if self.domain is not None:
             result['Domain'] = self.domain
+        if self.logstore is not None:
+            result['Logstore'] = self.logstore
+        if self.project is not None:
+            result['Project'] = self.project
+        if self.region is not None:
+            result['Region'] = self.region
         return result
 
     def from_map(self, m: dict = None):
         m = m or dict()
         if m.get('Domain') is not None:
             self.domain = m.get('Domain')
+        if m.get('Logstore') is not None:
+            self.logstore = m.get('Logstore')
+        if m.get('Project') is not None:
+            self.project = m.get('Project')
+        if m.get('Region') is not None:
+            self.region = m.get('Region')
         return self
 
 
@@ -30331,11 +30349,17 @@ class EnableRealtimeLogDeliveryRequest(TeaModel):
     def __init__(
         self,
         domain: str = None,
+        logstore: str = None,
+        project: str = None,
+        region: str = None,
     ):
         # The accelerated domain name for which you want to enable real-time log delivery. You can specify multiple domain names and separate them with commas (,).
         # 
         # This parameter is required.
         self.domain = domain
+        self.logstore = logstore
+        self.project = project
+        self.region = region
 
     def validate(self):
         pass
@@ -30348,12 +30372,24 @@ class EnableRealtimeLogDeliveryRequest(TeaModel):
         result = dict()
         if self.domain is not None:
             result['Domain'] = self.domain
+        if self.logstore is not None:
+            result['Logstore'] = self.logstore
+        if self.project is not None:
+            result['Project'] = self.project
+        if self.region is not None:
+            result['Region'] = self.region
         return result
 
     def from_map(self, m: dict = None):
         m = m or dict()
         if m.get('Domain') is not None:
             self.domain = m.get('Domain')
+        if m.get('Logstore') is not None:
+            self.logstore = m.get('Logstore')
+        if m.get('Project') is not None:
+            self.project = m.get('Project')
+        if m.get('Region') is not None:
+            self.region = m.get('Region')
         return self